admin/requirements.txt
django
djangorestframework
django-cors-headers
pika
python-dotenv
